Initial query rework: place and project with lookaside table.
