+++ dronefly/core/formatters/generic.py
@@ -32,7 +32,7 @@ from ..constants import (
     TRINOMIAL_ABBR,
     RANK_LEVELS,
 )
-from ..models.taxon_list import TaxonListMetadata
+from ..models import BaseCountFormatter, BaseFormatter, ListFormatter, TaxonListMetadata
 from ..utils import included_ranks
 from .constants import (
     ICONS,
@@ -484,27 +484,6 @@ def format_quality_grade(options: dict = {}):
         if needsid:
             adjectives.append("*Needs ID*")
     return adjectives
-
-
-class BaseFormatter:
-    def format():
-        raise NotImplementedError
-
-
-class ListFormatter(BaseFormatter):
-    def format_page():
-        raise NotImplementedError
-
-    def last_page():
-        raise NotImplementedError
-
-
-class BaseCountFormatter(BaseFormatter):
-    def count():
-        raise NotImplementedError
-
-    def description():
-        raise NotImplementedError
 
 
 class TaxonListFormatter(ListFormatter):
